[更新第三方库配置和代码重构]: 对项目配置和代码进行了一系列的更新与重构
- 更新了`3rdparty.pri`文件，移除了对静态链接Brotli库的引用，并统一了库的链接方式。
- 修改了多个源文件，将非const成员函数改为const，以支持在const上下文中调用。
- 更新了`colorutils.cc`文件，使用C++标准库中的`std::array`来定义颜色转换矩阵。
- 在`ffmpeg`库中，对`AVContextInfo`类进行了重构，增加了构造函数的显式声明，并在析构函数中释放`AVDictionary`。
- 更新了`CodecContext`类，增加了对音频和视频编码器选项的初始化，以及对编解码器上下文的查询和设置。
- 对`Frame`类进行了修改，增加了对帧持续时间的处理，并在复制帧属性时复制了持续时间。
- 更新了`Packet`类，增加了对包持续时间的处理，并改进了关键帧的判断逻辑。
- 在`player.cpp`中，为`processSpeedEvent`和`processVolumeEvent`函数添加了`static`和`const`限定符。
- 更新了`Subtitle`类，改进了字幕渲染和事件处理。
- 重构了`Ass`类，增加了对ASS库的错误消息回调，并改进了字幕事件的处理。
- 更新了`transcode.cc`文件，改进了音频和视频过滤器的初始化，并增加了对音频FIFO的处理。
- 更新了`videodecoder.cpp`和`videodisplay.cc`文件，改进了视频解码和显示逻辑。
- 在`videoframeconverter.cc`中，增加了对输入和输出像素格式的支持性检查。
- 更新了`OpenGLShaderProgram`和`ShaderUtils`类，改进了HDR和HLG颜色转换的处理。
- 重构了`VideoPreviewWidget`类，改进了视频预览任务的创建和执行。

+++ ffmpeg/event/seekevent.hpp
@@ -20,7 +20,7 @@ public:
     void setPosition(qint64 position) { m_position = position; }
     [[nodiscard]] auto position() const -> qint64 { return m_position; }
 
-    void setWaitCountdown(qint64 count) { m_latch.setCount(count); }
+    void setWaitCountdown(int count) { m_latch.setCount(count); }
     void countDown() { m_latch.countDown(); }
     void wait() { m_latch.wait(); }
 
@@ -36,7 +36,7 @@ public:
         : Event(parent)
         , m_relativePosition(relativePosition)
     {}
-    
+
     [[nodiscard]] auto type() const -> EventType override { return EventType::SeekRelative; }
 
     // second
